Remove vec allocation in write_next_block, add temp asserts
Right now there is a problem with the buffer size on Windows/WASAPI.
Despite requesting a supported BufferSize::Fixed, the output buffer has
an unexpected size. Some debug_assert!s have been added to catch this
for now. Ultimately, we might just have to deal with it and resize
occasionally.